Add files via upload
It is a simple game where a random pattern is generated and random buttons blink. The User has to remember the design and click on the buttons in the same way, on successful completion of each level, the difficulty increases and the game is reset whenever the user fails to replicate the shown pattern.